chore: enhance configuration with validation and CORS, update Jest settings for E2E tests
