+++ apps/api/test/jest-e2e.json
@@ -1,9 +1,23 @@
 {
   "moduleFileExtensions": ["js", "json", "ts"],
-  "rootDir": ".",
+  "rootDir": "..",
+  "roots": ["<rootDir>", "<rootDir>/../../packages/database/generated/prisma"],
   "testEnvironment": "node",
-  "testMatch": ["<rootDir>/test/e2e/**/*.e2e-spec.ts"],
+  "testMatch": ["<rootDir>/test/e2e/main-flows.e2e-spec.ts"],
+  "extensionsToTreatAsEsm": [".ts"],
+  "moduleNameMapper": {
+    "^@/(.*)$": "<rootDir>/src/$1",
+    "^@repo/db/prisma/(.*)$": "<rootDir>/../../packages/database/generated/prisma/$1.ts",
+    "^@dagrejs/dagre$": "<rootDir>/test/e2e/utils/dagre.mock.ts"
+  },
   "transform": {
-    "^.+\\.(t|j)s$": "ts-jest"
-  }
+    "^.+\\.ts$": [
+      "ts-jest",
+      {
+        "tsconfig": "<rootDir>/test/tsconfig.e2e.json",
+        "useESM": true
+      }
+    ]
+  },
+  "transformIgnorePatterns": ["[/\\\\]node_modules[/\\\\](?!@repo[/\\\\]db)"]
 }
